Refactor to remove hybrid trap density extraction from image_processing_functions. Restructured tests - all tests passed.
