+++ code/image_processing_functions.py
@@ -7,9 +7,6 @@ import numpy as np
 from scipy.optimize import fsolve
 from scipy import ndimage
 
-from . import data_fitting_functions
-
-
 
 """
 Returns the local variance in a 2d input array.
@@ -509,32 +506,6 @@ def get_polrot_images_from_atom_density(densities_1, densities_2, detuning_1A, d
 
 
 
-def get_hybrid_trap_densities_along_harmonic_axis(hybrid_trap_density_image, axicon_tilt_deg, axicon_diameter_pix, axicon_length_pix,
-                                                axicon_side_angle_deg, side_aspect_ratio,
-                                                um_per_pixel, center = None, rotate_data = True):
-    if(center is None):
-        center = data_fitting_functions.hybrid_trap_center_finder(hybrid_trap_density_image, axicon_tilt_deg, axicon_diameter_pix, axicon_length_pix)
-    if(rotate_data and axicon_tilt_deg != 0.0):
-        image_to_use, new_center = _rotate_and_crop_hybrid_image(hybrid_trap_density_image, center, axicon_tilt_deg)
-        center = new_center 
-    else:
-        image_to_use = hybrid_trap_density_image
-    x_center, y_center = center
-    hybrid_trap_radius_um = um_per_pixel * axicon_diameter_pix / 2.0
-    hybrid_trap_cross_sectional_area_um = get_hybrid_cross_section_um(hybrid_trap_radius_um, axicon_side_angle_deg, side_aspect_ratio)
-    radial_axis_index = 1
-    hybrid_trap_radial_integrated_density = um_per_pixel * np.sum(image_to_use, axis = radial_axis_index)
-    hybrid_trap_3D_density_harmonic_axis = hybrid_trap_radial_integrated_density / hybrid_trap_cross_sectional_area_um 
-    harmonic_axis_positions_um = um_per_pixel * (np.arange(len(image_to_use)) - y_center)
-    #Fit lorentzian to the 1D-integrated data to improve center-finding fidelity
-    results = data_fitting_functions.fit_lorentzian_with_offset(harmonic_axis_positions_um, hybrid_trap_3D_density_harmonic_axis)
-    popt, pcov = results
-    amp, center, gamma, offset = popt
-    refitted_harmonic_axis_positions_um = harmonic_axis_positions_um - center
-    return (refitted_harmonic_axis_positions_um, hybrid_trap_3D_density_harmonic_axis)
-
-
-
 """
 Given an image, return an angle to rotate it into the xy plane.
 
@@ -748,45 +719,6 @@ Kwargs as documented on https://pyabel.readthedocs.io/en/latest/abel.html
 """
 def inverse_abel(profile, **kwargs):
     return abel.Transform(profile, **kwargs).transform
-
-"""
-Convenience function for getting the actual areal cross section of the tilted oval of the hybrid trap.
-Note that the tilt angle is the angle made by the semimajor axis of the oval to the plane that the top imaging can see."""
-def get_hybrid_cross_section_um(top_radius_um, side_angle_deg, side_aspect_ratio):
-    side_angle_rad = side_angle_deg * np.pi / 180 
-    theta = side_angle_rad
-    semiminor_to_semimajor_ratio = 1.0 / side_aspect_ratio
-    s = semiminor_to_semimajor_ratio
-    #Slightly nontrivial geometry formula
-    seen_radius_to_semimajor_ratio = np.cos(theta) * np.sqrt(1 + np.square(s * np.tan(theta)))
-    semimajor_radius_um = top_radius_um / seen_radius_to_semimajor_ratio
-    semiminor_radius_um = semiminor_to_semimajor_ratio * semimajor_radius_um 
-    cross_section_um = np.pi * semimajor_radius_um * semiminor_radius_um
-    return cross_section_um
-
-
-def _rotate_and_crop_hybrid_image(image, center, rotation_angle_deg, x_crop_width = np.inf, y_crop_width = np.inf):
-    x_center, y_center = center 
-    image_y_width, image_x_width = image.shape
-    image_x_center = (image_x_width - 1) / 2.0 
-    image_y_center = (image_y_width - 1) / 2.0 
-    x_center_diff = x_center - image_x_center 
-    y_center_diff = y_center - image_y_center
-    rotation_angle_rad = np.pi / 180 * rotation_angle_deg
-    rotated_x_center_diff = np.cos(rotation_angle_rad) * x_center_diff + np.sin(rotation_angle_rad) * y_center_diff
-    rotated_y_center_diff = np.cos(rotation_angle_rad) * y_center_diff - np.sin(rotation_angle_rad) * x_center_diff 
-    rotated_x_center = image_x_center + rotated_x_center_diff 
-    rotated_y_center = image_y_center + rotated_y_center_diff 
-    rotated_image = ndimage.rotate(image, rotation_angle_deg, reshape = False)
-    cropped_y_max = int(min(image_y_width, np.round(rotated_y_center + y_crop_width / 2.0)))
-    cropped_y_min = int(max(0, np.round(rotated_y_center - y_crop_width / 2)))
-    cropped_x_max = int(min(image_x_width, np.round(rotated_x_center + x_crop_width / 2.0)))
-    cropped_x_min = int(max(0, np.round(rotated_x_center - x_crop_width / 2.0)))
-    cropped_rotated_image = rotated_image[cropped_y_min:cropped_y_max, cropped_x_min:cropped_x_max] 
-    final_x_center = rotated_x_center - cropped_x_min
-    final_y_center = rotated_y_center - cropped_y_min
-    return (cropped_rotated_image, (final_x_center, final_y_center)) 
-
 
 
 def get_saturation_counts_from_camera_parameters(pixel_length_at_atoms_m, imaging_time_s, camera_count_to_photon_factor, linewidth_Hz,
