+++ tests/test_analysis_functions.py
@@ -7,7 +7,7 @@ import shutil
 
 import astropy
 import numpy as np
-from scipy import ndimage
+import scipy
 
 #Temp import 
 import matplotlib.pyplot as plt
@@ -1038,6 +1038,66 @@ def test_get_hybrid_trap_densities_along_harmonic_axis():
         shutil.rmtree(measurement_pathname)
 
 
+def test_hybrid_trap_density_helper():
+    EXPECTED_X_CENTER = 228 
+    EXPECTED_Y_CENTER = 398
+    SAMPLE_UM_PER_PIXEL = 0.6
+    SAMPLE_AXICON_DIAMETER_PIX = 189
+    SAMPLE_AXICON_LENGTH_PIX = 250
+    SAMPLE_TILT_DEG = 6.3
+    SAMPLE_SIDE_TILT_DEG = 10
+    SAMPLE_SIDE_ASPECT_RATIO = 1.4
+    hybrid_trap_experiment_parameters = {
+        "top_um_per_pixel":SAMPLE_UM_PER_PIXEL, 
+        "axicon_diameter_pix":SAMPLE_AXICON_DIAMETER_PIX,
+        "axicon_tilt_deg":SAMPLE_TILT_DEG,
+        "axicon_side_aspect_ratio":SAMPLE_SIDE_ASPECT_RATIO, 
+        "axicon_side_angle_deg": SAMPLE_SIDE_TILT_DEG,
+        "hybrid_trap_typical_length_pix":SAMPLE_AXICON_LENGTH_PIX,
+    }
+    try:
+        measurement_pathname, my_measurement, _ = create_measurement("top_double", experiment_param_values = hybrid_trap_experiment_parameters)
+        sample_hybrid_trap_data = np.load('resources/Sample_Box_Exp.npy')
+        hybrid_trap_harmonic_positions, hybrid_trap_harmonic_data = analysis_functions._hybrid_trap_density_helper(my_measurement, sample_hybrid_trap_data)
+        filtered_hybrid_trap_harmonic_data = scipy.signal.savgol_filter(hybrid_trap_harmonic_data, 15, 2)
+        max_index = np.argmax(hybrid_trap_harmonic_data) 
+        max_value = hybrid_trap_harmonic_data[max_index]
+        CENTER_SNIPPET_HALF_WIDTH = 10
+        center_snippet = sample_hybrid_trap_data[EXPECTED_Y_CENTER - CENTER_SNIPPET_HALF_WIDTH:EXPECTED_Y_CENTER+CENTER_SNIPPET_HALF_WIDTH, 
+                                                EXPECTED_X_CENTER - CENTER_SNIPPET_HALF_WIDTH: EXPECTED_X_CENTER + CENTER_SNIPPET_HALF_WIDTH]
+        center_snippet_average_2d_density = np.sum(center_snippet) / center_snippet.size
+        center_snippet_average_3d_density = center_snippet_average_2d_density / (SAMPLE_UM_PER_PIXEL * SAMPLE_AXICON_DIAMETER_PIX)
+        max_index = np.argmax(filtered_hybrid_trap_harmonic_data)
+        max_value = filtered_hybrid_trap_harmonic_data[max_index] 
+        max_position = hybrid_trap_harmonic_positions[max_index] 
+        assert(np.abs(max_position) < 10) 
+        assert(np.abs((center_snippet_average_3d_density - max_value) / center_snippet_average_3d_density < 1e-1))
+    finally:
+        shutil.rmtree(measurement_pathname)
+
+
+def test_rotate_and_crop_hybrid_image():
+    X_SIZE = 300 
+    Y_SIZE = 300 
+    X_CENTER = 110 
+    Y_CENTER = 203 
+    GAUSSIAN_X_WIDTH = 50
+    GAUSSIAN_Y_WIDTH = 20
+    center = (X_CENTER, Y_CENTER)
+    y_indices, x_indices = np.mgrid[0:Y_SIZE, 0:X_SIZE]
+    gaussian_data = data_fitting_functions.two_dimensional_gaussian(x_indices, y_indices, 1.0, X_CENTER, Y_CENTER, GAUSSIAN_X_WIDTH, GAUSSIAN_Y_WIDTH, 0)
+    X_CROP_WIDTH = 100 
+    Y_CROP_WIDTH = 150
+    cropped_rotated_image, cropped_rotated_center = analysis_functions._rotate_and_crop_hybrid_image(gaussian_data, center, 90, 
+                                                                                x_crop_width = X_CROP_WIDTH, y_crop_width = Y_CROP_WIDTH)
+    rotated_x_center, rotated_y_center = cropped_rotated_center 
+    assert (np.abs(rotated_x_center - X_CROP_WIDTH / 2.0 < 1e-3))
+    assert (np.abs(rotated_y_center - Y_CROP_WIDTH / 2.0 < 1e-3))
+    EXPECTED_COUNTS_SUM = 5375.8749
+    counts_sum = np.sum(cropped_rotated_image)
+    assert (np.abs(counts_sum - EXPECTED_COUNTS_SUM) < 1e-3)
+
+
 def test_get_hybrid_trap_average_energy():
     hf_atom_density_experiment_param_values = {
         "state_1_unitarity_res_freq_MHz": 0.0,
@@ -1787,9 +1847,7 @@ def test_get_box_in_situ_fermi_energies_from_counts():
 
         box_radius_um = um_per_pixel * hf_atom_density_experiment_param_values["axicon_diameter_pix"] / 2
         box_length_um = um_per_pixel * hf_atom_density_experiment_param_values["box_length_pix"]
-        box_cross_section_um = image_processing_functions.get_hybrid_cross_section_um(box_radius_um, 
-                                                            hf_atom_density_experiment_param_values["axicon_side_angle_deg"], 
-                                                            hf_atom_density_experiment_param_values["axicon_side_aspect_ratio"])
+        box_cross_section_um = analysis_functions.get_hybrid_cross_section_um(my_measurement)
         expected_fermi_energy = science_functions.get_box_fermi_energy_from_counts(expected_atom_counts, box_cross_section_um, box_length_um)
         
         #Now get them from the analysis function...
@@ -1830,7 +1888,7 @@ def test_get_rapid_ramp_densities_along_harmonic_axis():
     }
     rotation_angle_deg = hf_atom_density_experiment_param_values["rr_tilt_deg"]
     default_absorption_image = get_default_absorption_image()
-    rotated_default_absorption_image = ndimage.rotate(default_absorption_image, -rotation_angle_deg, reshape = False)
+    rotated_default_absorption_image = scipy.ndimage.rotate(default_absorption_image, -rotation_angle_deg, reshape = False)
     rotated_default_image_stack = generate_image_stack_from_absorption(rotated_default_absorption_image)
     try:
         measurement_pathname, my_measurement, my_run = create_measurement("top_double", image_stack = rotated_default_image_stack, 
@@ -1841,7 +1899,7 @@ def test_get_rapid_ramp_densities_along_harmonic_axis():
         rr_density_pre_rotation = -np.log(rotated_default_absorption_image) / li_6_res_cross_section
         roi_xmin, roi_ymin, roi_xmax, roi_ymax = DEFAULT_ABSORPTION_IMAGE_ROI
         rr_density_pre_rotation_cropped = rr_density_pre_rotation[roi_ymin:roi_ymax, roi_xmin:roi_xmax]
-        rr_density_expected = ndimage.rotate(rr_density_pre_rotation_cropped, rotation_angle_deg, reshape = False)
+        rr_density_expected = scipy.ndimage.rotate(rr_density_pre_rotation_cropped, rotation_angle_deg, reshape = False)
         rr_integrated_density_expected = np.sum(rr_density_expected, axis = 1) * um_per_pixel
         #Test getting densities both by storing and by re-processing
         rr_integrated_densities = analysis_functions.get_rapid_ramp_densities_along_harmonic_axis(my_measurement, my_run, imaging_mode = "abs", 
@@ -1878,7 +1936,7 @@ def test_get_rr_condensate_fractions_fit():
     }
     rotation_angle_deg = hf_atom_density_experiment_param_values["rr_tilt_deg"]
     rr_condensate_sample_absorption_image = get_rr_condensate_sample_absorption_image()
-    rotated_rr_condensate_sample_absorption_image = ndimage.rotate(rr_condensate_sample_absorption_image, -rotation_angle_deg, reshape = False)
+    rotated_rr_condensate_sample_absorption_image = scipy.ndimage.rotate(rr_condensate_sample_absorption_image, -rotation_angle_deg, reshape = False)
     rotated_default_image_stack = generate_image_stack_from_absorption(rotated_rr_condensate_sample_absorption_image)
     try:
         measurement_pathname, my_measurement, my_run = create_measurement("top_double", image_stack = rotated_default_image_stack, 
@@ -1912,7 +1970,7 @@ def test_get_rr_condensate_fractions_box():
     }
     rotation_angle_deg = hf_atom_density_experiment_param_values["rr_tilt_deg"]
     rr_condensate_sample_absorption_image = get_rr_condensate_sample_absorption_image()
-    rotated_rr_condensate_sample_absorption_image = ndimage.rotate(rr_condensate_sample_absorption_image, -rotation_angle_deg, reshape = False)
+    rotated_rr_condensate_sample_absorption_image = scipy.ndimage.rotate(rr_condensate_sample_absorption_image, -rotation_angle_deg, reshape = False)
     rotated_default_image_stack = generate_image_stack_from_absorption(rotated_rr_condensate_sample_absorption_image)
     rr_roi = [100, 256 - RR_CONDENSATE_WIDTH, 400, 256 + RR_CONDENSATE_WIDTH]
     shifted_rr_roi = [100, 256 - 3*RR_CONDENSATE_WIDTH, 400, 256 - RR_CONDENSATE_WIDTH]
@@ -2076,6 +2134,17 @@ def test_get_saturation_counts_top():
         assert np.isclose(expected_saturation_counts_fudged, saturation_counts_fudged)
     finally:
         shutil.rmtree(measurement_pathname)
+
+
+def test_hybrid_cross_section_geometry_formula():
+    SAMPLE_SIDE_ANGLE_DEG = 45
+    SAMPLE_SIDE_ASPECT_RATIO = 2 
+    SAMPLE_TOP_RADIUS_UM = 100
+    #From hand-evaluation of the formula
+    EXPECTED_CROSS_SECTION_UM2 = 25132.74123
+    cross_section = analysis_functions._hybrid_cross_section_geometry_formula(
+        SAMPLE_TOP_RADIUS_UM, SAMPLE_SIDE_ANGLE_DEG, SAMPLE_SIDE_ASPECT_RATIO)
+    assert np.isclose(cross_section, EXPECTED_CROSS_SECTION_UM2)
